fix(ci): split development panel read model under economy file-size ceiling
Refs #4175
